requirements.txt
selene==2.0.0b14
pytest==7.2
allure-pytest>=2.12.0